+++ rage/src/bin/rage-mount/main.rs
@@ -3,6 +3,7 @@
 use age::{
     armor::ArmoredReader,
     cli_common::{read_identities, read_secret, StdinGuard},
+    scrypt,
     stream::StreamReader,
 };
 use clap::{CommandFactory, Parser};
@@ -210,12 +211,19 @@ fn main() -> Result<(), Error> {
     let mut stdin_guard = StdinGuard::new(false);
 
     match age::Decryptor::new_buffered(ArmoredReader::new(file))? {
-        age::Decryptor::Passphrase(decryptor) => {
+        age::Decryptor::Recipients(decryptor) if decryptor.is_scrypt() => {
             match read_secret(&fl!("type-passphrase"), &fl!("prompt-passphrase"), None) {
-                Ok(passphrase) => decryptor
-                    .decrypt(&passphrase, opts.max_work_factor)
-                    .map_err(|e| e.into())
-                    .and_then(|stream| mount_stream(stream, types, mountpoint)),
+                Ok(passphrase) => {
+                    let mut identity = scrypt::Identity::new(passphrase);
+                    if let Some(max_work_factor) = opts.max_work_factor {
+                        identity.set_max_work_factor(max_work_factor);
+                    }
+
+                    decryptor
+                        .decrypt(Some(&identity as _).into_iter())
+                        .map_err(|e| e.into())
+                        .and_then(|stream| mount_stream(stream, types, mountpoint))
+                }
                 Err(_) => Ok(()),
             }
         }